moved addAppointment button into appointments.tsx
