--- %ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,1 +1,0 @@
-1305.2405.14022.0
